Fixed constraints + minor adjustments
