--- iDad/IDadTableViewCell.xib
@@ -1,78 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="14490.70" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina6_1" orientation="portrait">
-        <adaptation id="fullscreen"/>
-    </device>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="14490.49"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <tableViewCell contentMode="scaleToFill" selectionStyle="default" indentationWidth="10" reuseIdentifier="IDadTableViewCell" rowHeight="252" id="KGk-i7-Jjw" customClass="IDadTableViewCell" customModule="iDad" customModuleProvider="target">
-            <rect key="frame" x="0.0" y="0.0" width="445" height="252"/>
-            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-            <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" tableViewCell="KGk-i7-Jjw" id="H2p-sc-9uM">
-                <rect key="frame" x="0.0" y="0.0" width="445" height="251.5"/>
-                <autoresizingMask key="autoresizingMask"/>
-                <subviews>
-                    <stackView opaque="NO" contentMode="scaleToFill" alignment="top" spacing="20" translatesAutoresizingMaskIntoConstraints="NO" id="tvu-9f-NL6">
-                        <rect key="frame" x="20" y="25" width="418" height="201"/>
-                        <subviews>
-                            <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="Bak-ho-o58">
-                                <rect key="frame" x="0.0" y="0.0" width="186" height="201"/>
-                                <constraints>
-                                    <constraint firstAttribute="width" constant="186" id="Z13-NS-uCU"/>
-                                    <constraint firstAttribute="height" constant="201" id="uf4-1d-69G"/>
-                                </constraints>
-                            </imageView>
-                            <stackView opaque="NO" contentMode="scaleToFill" axis="vertical" alignment="top" spacing="8" translatesAutoresizingMaskIntoConstraints="NO" id="SZ3-wW-NrD">
-                                <rect key="frame" x="206" y="0.0" width="212" height="181"/>
-                                <subviews>
-                                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="&quot;The meaning of life is found in the adoption of resposibility&quot;" textAlignment="natural" lineBreakMode="wordWrap" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontForContentSizeCategory="YES" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="sut-pU-Niu">
-                                        <rect key="frame" x="0.0" y="0.0" width="212" height="149"/>
-                                        <constraints>
-                                            <constraint firstAttribute="height" constant="149" id="eJk-nZ-s0V"/>
-                                            <constraint firstAttribute="width" constant="212" id="hLS-AO-Bqo"/>
-                                        </constraints>
-                                        <fontDescription key="fontDescription" type="system" pointSize="23"/>
-                                        <nil key="textColor"/>
-                                        <nil key="highlightedColor"/>
-                                    </label>
-                                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Jordan B. Peterson" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontForContentSizeCategory="YES" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="DVk-53-1oy">
-                                        <rect key="frame" x="0.0" y="157" width="177.5" height="24"/>
-                                        <fontDescription key="fontDescription" type="boldSystem" pointSize="20"/>
-                                        <nil key="textColor"/>
-                                        <nil key="highlightedColor"/>
-                                    </label>
-                                </subviews>
-                            </stackView>
-                        </subviews>
-                        <constraints>
-                            <constraint firstItem="SZ3-wW-NrD" firstAttribute="leading" secondItem="Bak-ho-o58" secondAttribute="trailing" constant="20" id="7wN-00-w5x"/>
-                            <constraint firstAttribute="bottom" secondItem="SZ3-wW-NrD" secondAttribute="bottom" constant="20" id="ELm-7b-G84"/>
-                            <constraint firstAttribute="trailing" secondItem="SZ3-wW-NrD" secondAttribute="trailing" id="hp7-8h-f07"/>
-                            <constraint firstItem="SZ3-wW-NrD" firstAttribute="top" secondItem="tvu-9f-NL6" secondAttribute="top" id="zD7-5Q-Nbn"/>
-                        </constraints>
-                    </stackView>
-                </subviews>
-                <constraints>
-                    <constraint firstAttribute="bottom" secondItem="tvu-9f-NL6" secondAttribute="bottom" constant="25.5" id="adl-Ja-aLI"/>
-                    <constraint firstItem="tvu-9f-NL6" firstAttribute="top" secondItem="H2p-sc-9uM" secondAttribute="top" constant="25" id="c2P-K5-t6g"/>
-                    <constraint firstAttribute="trailing" secondItem="tvu-9f-NL6" secondAttribute="trailing" constant="7" id="c4e-gI-H9z"/>
-                    <constraint firstItem="tvu-9f-NL6" firstAttribute="leading" secondItem="H2p-sc-9uM" secondAttribute="leading" constant="20" id="ilf-Do-ydB"/>
-                </constraints>
-            </tableViewCellContentView>
-            <viewLayoutGuide key="safeArea" id="njF-e1-oar"/>
-            <connections>
-                <outlet property="nameLabel" destination="DVk-53-1oy" id="xMr-v1-m30"/>
-                <outlet property="profileImageView" destination="Bak-ho-o58" id="KZu-PG-7m9"/>
-                <outlet property="quoteLabel" destination="sut-pU-Niu" id="99j-XR-Yif"/>
-            </connections>
-            <point key="canvasLocation" x="68.840579710144937" y="75"/>
-        </tableViewCell>
-    </objects>
-</document>
