+++ iDad/ViewControllers/IDadProfileViewController.swift
@@ -10,14 +10,14 @@ import UIKit
 
 class IDadProfileViewController: UIViewController {
 
-    var iDad: IDadViewModel? = nil
+    var iDadViewModel: IDadViewModel? = nil
     
     override func viewDidLoad() {
         super.viewDidLoad()
 
         // Do any additional setup after loading the view.
         view.backgroundColor = .blue
-        title = iDad?.name
+        title = iDadViewModel?.name
     }
