+++ iDad/Views/IDadTableViewCell.xib
@@ -1,0 +1,76 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="14490.70" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
+    <device id="retina6_1" orientation="portrait">
+        <adaptation id="fullscreen"/>
+    </device>
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="14490.49"/>
+        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <tableViewCell contentMode="scaleToFill" selectionStyle="default" indentationWidth="10" reuseIdentifier="IDadTableViewCell" rowHeight="239" id="KGk-i7-Jjw" customClass="IDadTableViewCell" customModule="iDad" customModuleProvider="target">
+            <rect key="frame" x="0.0" y="0.0" width="406" height="212"/>
+            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+            <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" tableViewCell="KGk-i7-Jjw" id="H2p-sc-9uM">
+                <rect key="frame" x="0.0" y="0.0" width="406" height="211.5"/>
+                <autoresizingMask key="autoresizingMask"/>
+                <subviews>
+                    <stackView opaque="NO" contentMode="scaleToFill" distribution="fillProportionally" alignment="top" spacing="18" translatesAutoresizingMaskIntoConstraints="NO" id="tvu-9f-NL6" userLabel="H Stack View">
+                        <rect key="frame" x="5" y="5" width="396" height="201.5"/>
+                        <subviews>
+                            <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" horizontalCompressionResistancePriority="749" verticalCompressionResistancePriority="749" translatesAutoresizingMaskIntoConstraints="NO" id="Bak-ho-o58">
+                                <rect key="frame" x="0.0" y="0.0" width="142" height="201"/>
+                                <constraints>
+                                    <constraint firstAttribute="width" priority="999" constant="186" id="Z13-NS-uCU"/>
+                                    <constraint firstAttribute="height" constant="201" id="uf4-1d-69G"/>
+                                </constraints>
+                            </imageView>
+                            <stackView opaque="NO" contentMode="scaleToFill" axis="vertical" alignment="top" spacing="-12" translatesAutoresizingMaskIntoConstraints="NO" id="SZ3-wW-NrD" userLabel="V Stack View">
+                                <rect key="frame" x="160" y="0.0" width="236" height="201"/>
+                                <subviews>
+                                    <label opaque="NO" userInteractionEnabled="NO" contentMode="center" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="&quot;The meaning of life is found in the adoption of resposibility&quot;" textAlignment="natural" lineBreakMode="wordWrap" numberOfLines="0" baselineAdjustment="alignBaselines" minimumScaleFactor="0.14999999999999999" adjustsFontForContentSizeCategory="YES" translatesAutoresizingMaskIntoConstraints="NO" id="sut-pU-Niu">
+                                        <rect key="frame" x="0.0" y="0.0" width="212" height="149"/>
+                                        <constraints>
+                                            <constraint firstAttribute="height" constant="149" id="eJk-nZ-s0V"/>
+                                            <constraint firstAttribute="width" constant="212" id="hLS-AO-Bqo"/>
+                                        </constraints>
+                                        <fontDescription key="fontDescription" type="system" pointSize="24"/>
+                                        <nil key="textColor"/>
+                                        <nil key="highlightedColor"/>
+                                    </label>
+                                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Jordan B. Peterson" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" minimumScaleFactor="0.14999999999999999" adjustsFontForContentSizeCategory="YES" translatesAutoresizingMaskIntoConstraints="NO" id="DVk-53-1oy">
+                                        <rect key="frame" x="0.0" y="137" width="236" height="64"/>
+                                        <fontDescription key="fontDescription" type="boldSystem" pointSize="28"/>
+                                        <nil key="textColor"/>
+                                        <nil key="highlightedColor"/>
+                                    </label>
+                                </subviews>
+                            </stackView>
+                        </subviews>
+                        <constraints>
+                            <constraint firstItem="Bak-ho-o58" firstAttribute="height" secondItem="SZ3-wW-NrD" secondAttribute="height" id="9LI-uQ-fW7"/>
+                            <constraint firstItem="Bak-ho-o58" firstAttribute="width" secondItem="SZ3-wW-NrD" secondAttribute="width" multiplier="0.6" id="axz-rW-sTD"/>
+                        </constraints>
+                    </stackView>
+                </subviews>
+                <constraints>
+                    <constraint firstItem="tvu-9f-NL6" firstAttribute="top" secondItem="H2p-sc-9uM" secondAttribute="top" constant="5" id="C0j-ME-Re2"/>
+                    <constraint firstItem="tvu-9f-NL6" firstAttribute="leading" secondItem="H2p-sc-9uM" secondAttribute="leading" constant="5" id="EFe-dZ-Sfa"/>
+                    <constraint firstAttribute="trailing" secondItem="tvu-9f-NL6" secondAttribute="trailing" constant="5" id="Ryy-aa-X1A"/>
+                    <constraint firstAttribute="bottom" secondItem="tvu-9f-NL6" secondAttribute="bottom" constant="5" id="hRT-qG-oZq"/>
+                </constraints>
+            </tableViewCellContentView>
+            <viewLayoutGuide key="safeArea" id="njF-e1-oar"/>
+            <connections>
+                <outlet property="nameLabel" destination="DVk-53-1oy" id="xMr-v1-m30"/>
+                <outlet property="profileImageView" destination="Bak-ho-o58" id="KZu-PG-7m9"/>
+                <outlet property="quoteLabel" destination="sut-pU-Niu" id="99j-XR-Yif"/>
+            </connections>
+            <point key="canvasLocation" x="152.89855072463769" y="134.26339285714286"/>
+        </tableViewCell>
+    </objects>
+</document>
